backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Bcrypt
PyJWT
requests
gunicorn
    